Fix worktree
Refactor the code
