+++ src/components/PokemonCard/index.tsx
@@ -10,10 +10,6 @@ type PokemonCardProps = {
 	name: string
 }
 
-type PokemonEvolutioCardProps = {
-	pokemon: Pokemon
-}
-
 const PokemonCard = ({ name }: PokemonCardProps) => {
 	const [pokemon, setPokemon] = useState<Pokemon>()
 	const [imageLoading, setImageLoading] = useState(true)
@@ -46,14 +42,6 @@ const PokemonCard = ({ name }: PokemonCardProps) => {
 									imageLoading ? {} : { display: 'block' }
 							}
 						/>
-						{
-							tooManyRequests ?
-								(
-									<h6 className='mx-auto'>
-										<span className='p-2 bg-red-500 mt-2'>Too many requests</span>
-									</h6>
-								) : null
-						}
 					</div>
 					<div className='w-full flex flex-col sm:items-start items-center sm:px-5 px-2 sm:py-2 py-1'>
 						<p className='text-xs font-bold text-[#888]'>Nº: {formatNumber(pokemon.id)}</p>
@@ -73,51 +61,6 @@ const PokemonCard = ({ name }: PokemonCardProps) => {
 		) : (
 			<div className='hidden' />
 		)
-	)
-}
-
-export const PokemonEvolutionCard = ({ pokemon }: PokemonEvolutioCardProps) => {
-	const [imageLoading, setImageLoading] = useState(true)
-	const [tooManyRequests, setTooManyRequests] = useState(false)
-
-	return (
-		<Link to={`/pokemon/${pokemon.id}`} state={{ fromApp: true, pokemon: pokemon }}>
-			<div className='w-full h-full cursor-pointer pb-2 hover:shadow duration-300 rounded-md p-4'>
-				<div className='w-full flex justify-center h-auto bg-[#f5f5f5] rounded-[50%] overflow-hidden'>
-					<img
-						src={pokemon?.sprites.front_default}
-						className='sm:w-full w-full h-48'
-						onLoad={() => setImageLoading(false)}
-						onError={() => setTooManyRequests(true)}
-						style={
-							tooManyRequests ? { display: 'none' } :
-								imageLoading ? {} : { display: 'block' }
-						}
-					/>
-					{
-						tooManyRequests ?
-							(
-								<h6 className='mx-auto'>
-									<span className='p-2 bg-red-500 mt-2'>Too many requests</span>
-								</h6>
-							) : null
-					}
-				</div>
-				<div className='w-full flex flex-col justify-center items-center sm:py-2 py-1'>
-					<p className='text-xs font-bold text-[#888]'>Nº: {formatNumber(pokemon.id)}</p>
-					<div className='mt-1'>
-						<p className='text-sm text-black'>{pokemon.name}</p>
-					</div>
-					<div className='flex sm:gap-2 md:gap-1 lg:gap-4 gap-1 mt-1'>
-						{
-							pokemon.types.map((t, i) => (
-								<TypeLabel key={i} name={t.type.name} />
-							))
-						}
-					</div>
-				</div>
-			</div>
-		</Link>
 	)
 }
