+++ src/pages/Detail/data/index.ts
@@ -1,0 +1,20 @@
+const PieChartColors = {
+	backgroundColor: [
+		'rgba(75, 192, 120, 0.2)',
+		'rgba(255, 99, 132, .2)',
+		'rgba(25, 88, 236, 0.2)',
+		'rgba(199, 102, 255, 0.2)',
+		'rgba(86, 255, 252, 0.2)',
+		'rgba(255, 159, 64, .2)',
+	],
+	borderColor: [
+		'rgba(75, 192, 120, 1)',
+		'rgba(255, 99, 132, 1)',
+		'rgba(25, 88, 236, 1)',
+		'rgba(199, 102, 255, 1)',
+		'rgba(86, 255, 252, 1)',
+		'rgba(255, 159, 64, 1)',
+	],
+}
+
+export default PieChartColors
